requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
bitarray==1.2.2
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cytoolz==0.11.2
discord-webhook==0.17.0
distlib==0.3.5
eth-abi==3.0.0
eth-account==0.6.1
eth-hash==0.3.3
eth-keyfile==0.6.0
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.1.0
eth-utils==2.0.0
filelock==3.7.1
Flask==2.2.2
frozenlist==1.3.0
hexbytes==0.2.2
idna==3.3
ipfshttpclient==0.8.0a2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.6.1
lru-dict==1.1.7
MarkupSafe==2.1.1
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
oauthlib==3.2.2
parsimonious==0.8.1
platformdirs==2.5.2
protobuf==3.20.1
pycryptodome==3.15.0
pyrsistent==0.18.1
python-dotenv==0.21.0
pywin32==304
requests==2.28.1
requests-oauthlib==1.3.1
rlp==3.0.0
six==1.16.0
toolz==0.11.2
tweepy==4.10.1
urllib3==1.26.9
varint==1.0.2
virtualenv==20.15.1
web3==6.0.0b3
websockets==10.3
Werkzeug==2.2.2
yarl==1.7.2
